Adding user interface
We correct the db.js file. Add the callback function to the db.all method

change all res.send() call to res.format() in which we render both json and html

we modify the articles.ejs and we strip the header and footer include.

There is an error when they are included.

We will open a bug issue for that.